+++ src/ListBroking/AppBundle/Service/Helper/FileHandlerServiceInterface.php
@@ -6,6 +6,7 @@
 namespace ListBroking\AppBundle\Service\Helper;
 
 use Doctrine\ORM\Query;
+use Exporter\Writer\WriterInterface;
 use Symfony\Component\Form\Form;
 use Symfony\Component\HttpFoundation\File\File;
 
@@ -31,26 +32,57 @@ interface FileHandlerServiceInterface
 
     /**
      * Generates a file using a Query object
-     * @param      $name
-     * @param      $extension
-     * @param      $query
-     * @param      $headers
-     * @param bool $zipped
+     * @param string     $name
+     * @param string     $extension
+     * @param Query      $query
+     * @param array      $headers
+     * @param bool       $zipped
      *
-     * @return mixed
+     * @return string
      */
     public function generateFileFromQuery($name, $extension, Query $query, $headers, $zipped = true);
+
 
     /**
      * Generates a file using an Array
      *
-     * @param      $name
-     * @param      $extension
-     * @param      $array
-     * @param bool $zipped
+     * @param string      $name
+     * @param string      $extension
+     * @param array       $array
+     * @param bool        $zipped
      *
-     * @return mixed
+     * @return string
      */
     public function generateFileFromArray ($name, $extension, $array, $zipped = true);
 
+    /**
+     * Generates a file Writer
+     *
+     * @param        $name
+     * @param string $extension
+     */
+    public function createFileWriter($name, $extension);
+
+    /**
+     * Iterates an array with a given Writer
+     *
+     * @param array           $array
+     */
+    public function writeArray($array);
+
+    /**
+     * Opens a given writer
+     *
+     * @return mixed
+     */
+    public function openWriter();
+
+    /**
+     * Closes a given Writer
+     *
+     * @param bool            $zipped
+     *
+     * @return string
+     */
+    public function closeWriter($zipped = true);
 }